refactor removed print satemants

+++ events/views.py
@@ -12,11 +12,17 @@ def events(request):
     context = {
         'events': events,
     }
+
     return render(request, 'events/events.html', context)
 
 
+@login_required
 def add_event(request):
     '''Admin can add new events to the database'''
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must be admin to add Events.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = EventForm(request.POST, request.FILES)
         if form.is_valid():
@@ -24,8 +30,9 @@ def add_event(request):
             messages.success(request, 'Successfully added event!')
             return redirect(reverse('events'))
         else:
-            messages.error(request, 
-                           'Failed to add event. Please ensure the form is valid.')
+            messages.error(request,
+                           'Failed to add event. \
+                            Please ensure the form is valid.')
     else:
         form = EventForm()
 
@@ -36,11 +43,15 @@ def add_event(request):
     return render(request, 'events/add_events.html', context)
 
 
+@login_required()
 def edit_event(request, event_id):
     """Allows admin to edit events """
-    print('edit', event_id)
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must be admin to edit Events.')
+        return redirect(reverse('home'))
     event = get_object_or_404(Events, pk=event_id)
-    print(event)
+
     if request.method == 'POST':
         form = EventForm(request.POST, request.FILES, instance=event)
         if form.is_valid():
@@ -50,7 +61,8 @@ def edit_event(request, event_id):
             return redirect(reverse('events'))
         else:
             messages.error(request,
-                           'Failed to update Events. Please ensure the form is valid.')
+                           'Failed to update Events.\
+                            Please ensure the form is valid.')
     else:
         form = EventForm(instance=event)
         messages.info(request, f'You are editing {event.event_name}')
@@ -63,13 +75,19 @@ def edit_event(request, event_id):
     return render(request, 'events/edit_event.html', context)
 
 
+@login_required()
 def delete_event(request, event_id):
-    print('delete')
+    """ Allow admin to delete events from the page """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must be admin to add Events.')
+        return redirect(reverse('home'))
+
     event = get_object_or_404(Events, pk=event_id)
     event.delete()
     events = Events.objects.all()
     context = {
         'events': events
     }
-    return render(request, 'events/events.html', context)
 
+    return render(request, 'events/events.html', context)
